directory [panframe]
 class caputretool の削除
 class SimplePrayerActivityにキャプチャー機能実装（途中）
